Refactored 'images' command (#80)
