+++ pkg/cli/images.go
@@ -5,18 +5,8 @@
 package cli
 
 import (
-	"fmt"
-	"os"
-	"sort"
-	"sync"
-	"text/tabwriter"
-	"time"
-
-	log "github.com/scaleway/scaleway-cli/vendor/github.com/Sirupsen/logrus"
-	"github.com/scaleway/scaleway-cli/vendor/github.com/docker/docker/pkg/units"
-
-	"github.com/scaleway/scaleway-cli/pkg/api"
-	"github.com/scaleway/scaleway-cli/pkg/utils"
+	"github.com/scaleway/scaleway-cli/pkg/commands"
+	"github.com/scaleway/scaleway-cli/vendor/github.com/Sirupsen/logrus"
 )
 
 var cmdImages = &Command{
@@ -39,159 +29,22 @@ var imagesQ bool       // -q flag
 var imagesNoTrunc bool // -no-trunc flag
 var imagesHelp bool    // -h, --help flag
 
-func runImages(cmd *Command, args []string) {
+func runImages(cmd *Command, rawArgs []string) {
 	if imagesHelp {
 		cmd.PrintUsage()
 	}
-	if len(args) != 0 {
+	if len(rawArgs) != 0 {
 		cmd.PrintShortUsage()
 	}
 
-	wg := sync.WaitGroup{}
-	chEntries := make(chan api.ScalewayImageInterface)
-	var entries = []api.ScalewayImageInterface{}
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		images, err := cmd.API.GetImages()
-		if err != nil {
-			log.Fatalf("unable to fetch images from the Scaleway API: %v", err)
-		}
-		for _, val := range *images {
-			creationDate, err := time.Parse("2006-01-02T15:04:05.000000+00:00", val.CreationDate)
-			if err != nil {
-				log.Fatalf("unable to parse creation date from the Scaleway API: %v", err)
-			}
-			chEntries <- api.ScalewayImageInterface{
-				Type:         "image",
-				CreationDate: creationDate,
-				Identifier:   val.Identifier,
-				Name:         val.Name,
-				Public:       val.Public,
-				Tag:          "latest",
-				VirtualSize:  float64(val.RootVolume.Size),
-			}
-		}
-	}()
-
-	if imagesA {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			snapshots, err := cmd.API.GetSnapshots()
-			if err != nil {
-				log.Fatalf("unable to fetch snapshots from the Scaleway API: %v", err)
-			}
-			for _, val := range *snapshots {
-				creationDate, err := time.Parse("2006-01-02T15:04:05.000000+00:00", val.CreationDate)
-				if err != nil {
-					log.Fatalf("unable to parse creation date from the Scaleway API: %v", err)
-				}
-				chEntries <- api.ScalewayImageInterface{
-					Type:         "snapshot",
-					CreationDate: creationDate,
-					Identifier:   val.Identifier,
-					Name:         val.Name,
-					Tag:          "<snapshot>",
-					VirtualSize:  float64(val.Size),
-					Public:       false,
-				}
-			}
-		}()
-
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			bootscripts, err := cmd.API.GetBootscripts()
-			if err != nil {
-				log.Fatalf("unable to fetch bootscripts from the Scaleway API: %v", err)
-			}
-			for _, val := range *bootscripts {
-				chEntries <- api.ScalewayImageInterface{
-					Type:       "bootscript",
-					Identifier: val.Identifier,
-					Name:       val.Title,
-					Tag:        "<bootscript>",
-					Public:     false,
-				}
-			}
-		}()
-
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			volumes, err := cmd.API.GetVolumes()
-			if err != nil {
-				log.Fatalf("unable to fetch volumes from the Scaleway API: %v", err)
-			}
-			for _, val := range *volumes {
-				creationDate, err := time.Parse("2006-01-02T15:04:05.000000+00:00", val.CreationDate)
-				if err != nil {
-					log.Fatalf("unable to parse creation date from the Scaleway API: %v", err)
-				}
-				chEntries <- api.ScalewayImageInterface{
-					Type:         "volume",
-					CreationDate: creationDate,
-					Identifier:   val.Identifier,
-					Name:         val.Name,
-					Tag:          "<volume>",
-					VirtualSize:  float64(val.Size),
-					Public:       false,
-				}
-			}
-		}()
+	args := commands.ImagesArgs{
+		All:     imagesA,
+		Quiet:   imagesQ,
+		NoTrunc: imagesNoTrunc,
 	}
-
-	go func() {
-		wg.Wait()
-		close(chEntries)
-	}()
-
-	done := false
-	for {
-		select {
-		case entry, ok := <-chEntries:
-			if !ok {
-				done = true
-				break
-			}
-			entries = append(entries, entry)
-		}
-		if done {
-			break
-		}
-	}
-
-	w := tabwriter.NewWriter(os.Stdout, 20, 1, 3, ' ', 0)
-	defer w.Flush()
-	if !imagesQ {
-		fmt.Fprintf(w, "REPOSITORY\tTAG\tIMAGE ID\tCREATED\tVIRTUAL SIZE\n")
-	}
-	sort.Sort(api.ByCreationDate(entries))
-	for _, image := range entries {
-		if imagesQ {
-			fmt.Fprintf(w, "%s\n", image.Identifier)
-		} else {
-			tag := image.Tag
-			shortID := utils.TruncIf(image.Identifier, 8, !imagesNoTrunc)
-			name := utils.Wordify(image.Name)
-			if !image.Public && image.Type == "image" {
-				name = "user/" + name
-			}
-			shortName := utils.TruncIf(name, 25, !imagesNoTrunc)
-			var creationDate, virtualSize string
-			if image.CreationDate.IsZero() {
-				creationDate = "n/a"
-			} else {
-				creationDate = units.HumanDuration(time.Now().UTC().Sub(image.CreationDate))
-			}
-			if image.VirtualSize == 0 {
-				virtualSize = "n/a"
-			} else {
-				virtualSize = units.HumanSize(image.VirtualSize)
-			}
-			fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%s\n", shortName, tag, shortID, creationDate, virtualSize)
-		}
+	ctx := cmd.GetContext(rawArgs)
+	err := commands.RunImages(ctx, args)
+	if err != nil {
+		logrus.Fatalf("Cannot execute 'images': %v", err)
 	}
 }
